Create wrapper around Parser::ParseDeclarationSpecifiers
This wrapper should consume any split tokens and set the presence condition of the current scope to match the first token that is not a split token. The idea is that this parser method is called when parsing any kind of declaration so we won't need to do this in multiple places.