Push to VariabilityStack in HandleIfDirective
This allows the correct presence condition to be assigned to tokens inside "#if defined" blocks.

The VariabilityLocation struct was modified. The "name" field was removed and replaced by a field called "condition" which stores a PresenceCondition pointer.

The method PresenceCondition::GetList was modified to take a vector of PresenceCondition pointers instead of a vector of macro names as strings.