Modify Preprocessor::HandleIfDirective to use EraseCachedTokens and LastCachedTokenRange properly
Before this commit, an assert would sometimes fail since LastCachedTokenRange was called when backtracking was not enabled. Based on usage in Pragma.cpp in the LexingFor_PragmaRAII destructor, LastCachedTokenRange should be called immediately before CommitBacktrackedTokens or Backtrack is called, and EraseCachedTokens should be called immediately after CommitBacktrackedTokens or Backtrack is called.