Handle special case where macro expansion is only 1 token long
`Preprocessor::HandleMacroExpandedIdentifier` was modified to assign a new presence condition to the macro expansion token to ensure it has the presence condition of the expansion location and not the definition location. Also see previous commit.

Created new helper method `Preprocessor::ComputeConditional` to do some of the work that `AssignConditional` was doing which would have been duplicated in `HandleMacroExpandedIdentifier`.