Disable inserting of split token based on conjunction
Previously, when disabling the inserting of a split token, the preprocessor was checking whether or not the current preprocessor token was equivalent to the new condition introduced by the #if/#ifdef. It *should've* been checking the current preprocessor condition against the conjunction of the current preprocessor condition AND the presence condition introduced by the #if/#ifdef. This will catch more cases where a split shouldn't be inserted.